# dashboardapi/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the api for users",
    "version": "version"
  },
  "tags": [{ "name": "PUBLIC ROUTES" }, { "description": "Public Route" }],
  "paths": {
    "/health": {
      "Summary": "Testing Health Check",
      "get": {
        "Description": "Testing Health Check for app",
        "tags": ["PUBLIC ROUTES"],
        "responses": {
          "200": {
            "description": "Health Check passed successfully"
          }
        }
      }
    },
    "/users": {
      "Summary": "Get All Users",
      "get": {
        "Description": "Get All Users",
        "tags": ["PUBLIC ROUTES"],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string"
            },
            "description": "Choose a role as admin or user"
          }
        ],
        "responses": {
          "200": {
            "description": "All Users Displayed"
          }
        }
      }
    },
    "/user/{id}": {
      "Summary": "Get All Users",
      "get": {
        "Description": "Get All Users",
        "tags": ["PUBLIC ROUTES"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Enter object id"
          }
        ],
        "responses": {
          "200": {
            "description": "get User Details"
          }
        }
      }
    },
    "/addUser": {
      "Summary": "Get All Users",
      "post": {
        "Description": "Get All Users",
        "tags": ["PUBLIC ROUTES"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "id": "string"
              },
              "example": {
                "name": "Keerti",
                "city": "Amsterdam",
                "phone": "989878462",
                "role": "User",
                "isActive": "true"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "get User Details"
          }
        }
      }
    },
    "/updateuser": {
      "Summary": "Update User",
      "put": {
        "Description": "Update User",
        "tags": ["PUBLIC ROUTES"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "id": "string"
              },
              "example": {
                "_id": "",
                "name": "Keerti",
                "city": "Amsterdam",
                "phone": "989878462",
                "role": "User",
                "isActive": "true"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "get User Details"
          }
        }
      }
    }
  }
}